Turn command_utils into a package
The single module already holds two unrelated halves: helpers that only
compute strings, and the entry points that actually launch a run. Give
each half its own module and keep the old import path working through the
package, so the coming backend objects have somewhere to land.
