+++ tests/fast/utils/command_recorder.py
@@ -1,4 +1,4 @@
-import miles.utils.external_utils.command_utils as command_utils
+from miles.utils.external_utils.command_utils import base_backend
 
 
 def record_commands(monkeypatch) -> list[str]:
@@ -15,8 +15,8 @@ def record_commands(monkeypatch) -> list[str]:
         commands.append(f"[multi_node num_nodes={num_nodes}] {cmd}")
         return ["0"]
 
-    monkeypatch.setattr(command_utils, "exec_command_cpu", fake_exec_command)
-    monkeypatch.setattr(command_utils, "exec_command_gpu", fake_exec_command)
-    monkeypatch.setattr(command_utils, "exec_command_multi_node", fake_exec_command_multi_node)
+    monkeypatch.setattr(base_backend, "exec_command_cpu", fake_exec_command)
+    monkeypatch.setattr(base_backend, "exec_command_gpu", fake_exec_command)
+    monkeypatch.setattr(base_backend, "exec_command_multi_node", fake_exec_command_multi_node)
 
     return commands
